+++ frontend/src/components/ErrorHandlerPages/LoggedIn/index.js
@@ -1,0 +1,1 @@
+export * from "./LoggedIn";
