+++ frontend/src/components/ErrorHandlerPages/LoggedIn/logged-in.module.scss
@@ -1,0 +1,36 @@
+.not-found {
+  background-color: #f7f7f7;
+}
+
+.error {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 90vh;
+  margin: 0 15% 3rem 15%;
+}
+
+.error-logo {
+  width: 100%;
+  height: auto;
+  object-fit: cover;
+}
+
+.error-content {
+  color: black;
+  font-size: 1.2rem;
+  margin-bottom: 1.5rem;
+}
+
+.error-text {
+  font-size: 2rem;
+  font-weight: 900;
+  color: var(--secondary-color);
+}
+
+@media screen and (max-width: 750px) {
+  .error {
+    margin: 0 5% 3rem 5%;
+  }
+}
