+++ frontend/src/pages/Login/Login.jsx
@@ -1,8 +1,6 @@
 import React, { useState, useRef } from "react";
 import { Button2 } from "../../components/util/Button/index";
 import style from "./login.module.scss";
-import { useDispatch } from "react-redux";
-import * as actions from "../../store/actions/actions";
 
 export function Login() {
   const [hidePassword, setHidePassword] = useState(false);
@@ -14,8 +12,6 @@ export function Login() {
   const handleChange = (e) => {
     setCredential({ ...credential, [e.target.name]: e.target.value });
   };
-
-  const dispatch = useDispatch();
 
   function loginUser(e) {
     e.preventDefault();
@@ -30,11 +26,6 @@ export function Login() {
         if (response.status === 200) {
           localStorage.setItem("token", res.token);
           localStorage.setItem("isSuperAdmin", res.isSuperAdmin);
-          dispatch({
-            type: actions.LOG_IN,
-            token: res.token,
-            super: res.isSuperAdmin,
-          });
           window.location = "/dashboard";
         } else {
           setStatus("Unauthorised");
